backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django backend for the personal website"
authors = [{ name = "Md. Mushfiqur Rahman", email = "mushfiqur.rahman.robin@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.0",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "pathspec>=0.12.1",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-decouple>=3.8",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.0",
]

[tool.ruff]
line-length = 112
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "double"

[tool.black]
line-length = 112

[tool.isort]
profile = "black"
line_length = 112
